Save values in select2 form field upon validation error
When editing a task or upon a validation error, the select2 form fields
will keep their selected values.
This is done kind of awkardly when editing a task since django gives the
values either as string or int array. That means that the cost of saving
the values is currently O(n^3)

Closes #50